pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["time9v"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "6.0.1"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gendiff:main'

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^7.4.2"
pytest-cov = "4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
